rowblk: move 64-bit tests build flag
Move tests that should only run on 64-bit architectures behind a build flag.
